Refactored Landing page stuff
