+++ app/src/layout/LandingPage/OptionsBar.css
@@ -1,0 +1,5 @@
+.landing-page .options-bar {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+}
